requirements.txt
beautifulsoup4==4.11.2
elasticsearch==8.17.0
Flask==2.2.2
protobuf==4.21.12
Requests==2.32.3
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
